+++ src/test/resources/HiveRunnerExtensionTest/test_query.sql
@@ -1,0 +1,7 @@
+CREATE DATABASE testdb;
+
+CREATE EXTERNAL TABLE testdb.test_table
+(
+  field1 string,
+  field2 string
+)